convert strings to/from bytes where necessary
The old (python 2.x compatible) version of pyOSC used strings for wire
data as well as API arguments. Strings in python 3 are sequences of
unicode characters, not bytes, so we have to use 'bytes' objects to send
data over wire, but we still expect string in API arguments. Hopefully,
this patch handles everything (at least the test seems to work).