Shortcut for simple message creation
Now you can provide the payload to a message via a constructor. The tail
of the arguments follows exactly the same convention as .append(...)
method, with which I'm not quite happy (multiple message arguments need
to be passed in an interable, just because sometimes you would want to
supply the typehint). We should probably add a new method for supplying
the payload with a better convention and then make the constructor
follow the new one. We can then phase out the old .append().